pyproject.toml
[project]
name = "vcspull"
version = "1.34.0"
description = "Manage and sync multiple git, mercurial, and svn repos"
requires-python = ">=3.9,<4.0"
license = { text = "MIT" }
authors = [
  {name = "Tony Narlock", email = "tony@git-pull.com"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Environment :: Web Environment",
  "Intended Audience :: Developers",
  "Operating System :: POSIX",
  "Operating System :: MacOS :: MacOS X",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Utilities",
  "Topic :: System :: Shells",
]
packages = [
  { include = "*", from = "src" },
]
include = [
  { path = "tests", format = "sdist" },
  { path = ".tmuxp.yaml", format = "sdist" },
  { path = "docs", format = "sdist" },
  { path = "examples", format = "sdist" },
  { path = "conftest.py", format = "sdist" },
  { path = "src/vcspull/py.typed" },
]
readme = 'README.md'
keywords = [
  "vcspull",
  "vcs",
  "git",
  "svn",
  "subversion",
  "hg",
  "mercurial",
  "manage",
  "manager",
  "sync",
  "fetcher",
  "updater",
  "json",
  "yaml",
]
homepage = "https://vcspull.git-pull.com"
dependencies = [
  "libvcs~=0.35.0",
  "colorama>=0.3.9",
  "PyYAML>=6.0",
  "pydantic>=2.10.6",
]

[project-urls]
"Bug Tracker" = "https://github.com/vcs-python/vcspull/issues"
Documentation = "https://vcspull.git-pull.com"
Repository = "https://github.com/vcs-python/vcspull"
Changes = "https://github.com/vcs-python/vcspull/blob/master/CHANGES"

[project.scripts]
vcspull = "vcspull:cli.cli"

[tool.uv]
dev-dependencies = [
  # Docs
  "sphinx",
  "furo",
  "gp-libs",
  "sphinx-autobuild",
  "sphinx-autodoc-typehints",
  "sphinx-inline-tabs",
  "sphinxext-opengraph",
  "sphinx-copybutton",
  "sphinxext-rediraffe",
  "sphinx-argparse",
  "autodoc_pydantic",
  "myst-parser",
  "linkify-it-py",
  # Testing
  "gp-libs",
  "pytest",
  "pytest-rerunfailures",
  "pytest-mock",
  "pytest-watcher",
  "hypothesis",
  # Coverage
  "codecov",
  "coverage",
  "pytest-cov",
  # Lint
  "ruff",
  "mypy",
  # Annotations
  "types-requests",
  "types-PyYAML",
  "types-colorama"
]

[dependency-groups]
docs = [
  "sphinx",
  "furo",
  "gp-libs",
  "sphinx-autobuild",
  "sphinx-autodoc-typehints",
  "sphinx-inline-tabs",
  "sphinxext-opengraph",
  "sphinx-copybutton",
  "sphinxext-rediraffe",
  "sphinx-argparse",
  "autodoc_pydantic",
  "myst-parser",
  "linkify-it-py",
]
testing = [
  "gp-libs",
  "pytest",
  "pytest-rerunfailures",
  "pytest-mock",
  "pytest-watcher",
  "hypothesis",
]
coverage =[
  "codecov",
  "coverage",
  "pytest-cov",
]
lint = [
  "ruff",
  "mypy",
]
typings = [
  "types-requests",
  "types-PyYAML",
  "types-colorama"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = 3.9
warn_unused_configs = true
files = [
  "src",
  "tests",
]
exclude = [
  "examples/",
  "scripts/",
]
strict = true

[[tool.mypy.overrides]]
module = [
  "shtab",
]
ignore_missing_imports = true

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "*/_*",
  "*/_compat.py",
  "docs/conf.py",
  "tests/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise NotImplementedError",
  "return NotImplemented",
  "def parse_args",
  "if TYPE_CHECKING:",
  "if t.TYPE_CHECKING:",
  "@overload( |$)",
  "from __future__ import annotations",
]

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
select = [
  "E", # pycodestyle
  "F", # pyflakes
  "I", # isort
  "UP", # pyupgrade
  "A", # flake8-builtins
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "COM", # flake8-commas
  "EM", # flake8-errmsg
  "Q", # flake8-quotes
  "PTH", # flake8-use-pathlib
  "SIM", # flake8-simplify
  "TRY", # Trycertatops
  "PERF", # Perflint
  "RUF", # Ruff-specific rules
  "D", # pydocstyle
  "FA100",  # future annotations
]
ignore = [
  "COM812", # missing trailing comma, ruff format conflict
]
extend-safe-fixes = [
  "UP006",
  "UP007",
]
pyupgrade.keep-runtime-typing = false

[tool.ruff.lint.flake8-builtins]
builtins-allowed-modules = [
  "dataclasses",
  "types",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
known-first-party = [
  "vcspull",
]
combine-as-imports = true
required-imports = [
  "from __future__ import annotations",
]

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["F401"]

[tool.pytest.ini_options]
addopts = "--tb=short --no-header --showlocals"
testpaths = [
  "src/vcspull",
  "tests",
  "docs",
]
filterwarnings = [
  "ignore:The frontend.Option(Parser)? class.*:DeprecationWarning::",
]
